+++ packages/knowledge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vendoai/knowledge
 
+## 0.36.3
+
+### Patch Changes
+
+- @vendoai/core@0.36.3
+
+## 0.36.2
+
+### Patch Changes
+
+- @vendoai/core@0.36.2
+
 ## 0.36.1
 
 ### Patch Changes