+++ packages/vendoai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # vendoai
 
+## 0.36.3
+
+### Patch Changes
+
+- @vendoai/vendo@0.36.3
+
+## 0.36.2
+
+### Patch Changes
+
+- Updated dependencies [66cf10a]
+  - @vendoai/vendo@0.36.2
+
 ## 0.36.1
 
 ### Patch Changes